// src/lib/api/generated/model/notificationResDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BE SM Service API
 * BE SM Service API
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationResDtoEvent } from './notificationResDtoEvent';

export interface NotificationResDto {
  id: string;
  isRead: boolean;
  event: NotificationResDtoEvent;
  pageUrl: string;
  description: string;
  readonly createdAt: string;
}